openfeature_project/requirements.txt
anyio==4.4.0
asgiref==3.8.1
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
Django==5.0.6
djangorestframework==3.15.1
grpcio==1.64.1
h11==0.14.0
harness-featureflags==1.6.1
httpcore==1.0.5
httpx==0.27.0
idna==3.7
mmh3==4.1.0
openfeature-provider-flagd==0.1.5
openfeature_sdk==0.7.0
panzi-json-logic==1.0.1
protobuf==5.27.1
PyJWT==2.8.0
PyYAML==6.0.1
requests==2.32.3
semver==3.0.2
sniffio==1.3.1
sqlparse==0.5.0
tenacity==8.2.2
urllib3==2.2.2
